Move flake8-print violations
