+++ src/rules/flake8_print/rules/mod.rs
@@ -1,3 +1,3 @@
-pub use print_call::print_call;
+pub use print_call::{print_call, PPrintFound, PrintFound};
 
 mod print_call;
