+++ src/rules/flake8_print/rules/print_call.rs
@@ -1,12 +1,43 @@
 use log::error;
 use rustpython_ast::{Expr, Keyword, Stmt, StmtKind};
 
+use ruff_macros::derive_message_formats;
+
 use crate::ast::helpers::is_const_none;
 use crate::ast::types::Range;
 use crate::autofix::helpers;
 use crate::checkers::ast::Checker;
+use crate::define_violation;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::AlwaysAutofixableViolation;
+
+define_violation!(
+    pub struct PrintFound;
+);
+impl AlwaysAutofixableViolation for PrintFound {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("`print` found")
+    }
+
+    fn autofix_title(&self) -> String {
+        "Remove `print`".to_string()
+    }
+}
+
+define_violation!(
+    pub struct PPrintFound;
+);
+impl AlwaysAutofixableViolation for PPrintFound {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("`pprint` found")
+    }
+
+    fn autofix_title(&self) -> String {
+        "Remove `pprint`".to_string()
+    }
+}
 
 /// T201, T203
 pub fn print_call(checker: &mut Checker, func: &Expr, keywords: &[Keyword]) {
@@ -34,11 +65,11 @@ pub fn print_call(checker: &mut Checker, func: &Expr, keywords: &[Keyword]) {
                     }
                 }
             }
-            Diagnostic::new(violations::PrintFound, Range::from_located(func))
+            Diagnostic::new(PrintFound, Range::from_located(func))
         } else if call_path.as_ref().map_or(false, |call_path| {
             *call_path.as_slice() == ["pprint", "pprint"]
         }) {
-            Diagnostic::new(violations::PPrintFound, Range::from_located(func))
+            Diagnostic::new(PPrintFound, Range::from_located(func))
         } else {
             return;
         }
